Change wire protocol between the microcontrollers
In order to eliminate flicker when enabling twinkle modes, we want to
adjust colors atomically.  In order to preserve position within a video
when switching directories, we want to set the position atomically with
the directory and video change.  In order to support these, switch our
wire protocol to have only 3 commands: change colors, change speeds, and
change directory/video/position.