Reduce max size of DNS packets to accept
Given our memory size constraints and considering heap fragmentation,
the amount we were allocating was on the large side, and I saw at least
one MemoryError raised here because of it.  It doesn't help that
MicroPython preallocates a byte array the size of whatever max size you
pass to recvfrom, regardless of the actual size of the payload.

Note that 512 bytes is also the chunk size that we read index.html in.