Correct handling of set-position without set-video
Previously we would incorrectly re-send the last video that was sent to
us via json to the Arduino, which could reset it to an older video if it
had already reached the end and advanced on its own.  Instead, set the
position in the last video the Arduino told us it was on, unless the UI
tells us otherwise.